Отделил UI от логики: controllers + nav_bar В папке pages для goals и home остался только UI
