+++ fincontrolapp/components/__init__.py
@@ -1,5 +1,6 @@
 from .base_page import BasePage
 from .theme import AppTheme
 from .dialogs import show_dialog, close_dialog
+from .nav_bar import build_nav
 
-__all__ = ["BasePage", "AppTheme", "show_dialog", "close_dialog"]
+__all__ = ["BasePage", "AppTheme", "show_dialog", "close_dialog", "build_nav"]
