+++ fincontrolapp/controllers/__init__.py
@@ -1,0 +1,19 @@
+# from .auth_ctrl import AuthController
+# from .expenses_ctrl import ExpensesController
+from .goals_ctrl import GoalsController
+from .home_ctrl import HomeController
+# from .income_ctrl import IncomeController
+# from .settings_ctrl import SettingsController
+# from .subscriptions_ctrl import SubscriptionsController
+# from .transactions_ctrl import TransactionsController
+
+__all__ = [
+    # "AuthController",
+    # "ExpensesController",
+    "GoalsController",
+    "HomeController",
+    # "IncomeController",
+    # "SettingsController",
+    # "SubscriptionsController",    
+    # "TransactionsController",
+]
